requirements.txt
python-telegram-bot==13.15
pymongo==4.7.0
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.2
werkzeug==3.0.0
gunicorn==20.1.0